Harden template type annotations in DOMDocumentCaster (#30)
The second parameter of `DOMDocumentCaster::cast()`
must be a string, whereas its annotation was loosely
defined as `T`. However, the type of `T` is not
explicitly defined, leading to issues with static
code analyzers, e.g. PHPStan. To resolve this issue,
`T` is now defined as subtype of `object` and the
second parameter of `DOMDocumentCaster::cast()` is
described as `class-string` of `T`.